fix: align Avalonia fields with WinForms for WU2-WU6 forms
- ItemFE6: expand from 10 to 22 fields with full write-back
- UnitFE6: expand with all WinForms fields
- MapSetting: complete rewrite with all D/W/B/P fields + write-back
- MoveCostEditor: add all 64 terrain cost fields
- Support forms (Unit/FE6, Talk/FE6/FE7): complete field coverage
- SoundRoom: fix D0/D12 read widths (u16→u32) to match WinForms
- SongTable: add D4 field
- StatusUnitsMenu: add D8 field
- ED: add B2/B3 fields
- SupportAttribute: add B7 field
- Fix P/D equivalence in field completeness test
- Update test expectations for new window sizes and ROM offsets

All 1460 tests pass. Full solution builds clean.

Prompt: Continue WinForms vs Avalonia form comparison & alignment plan

+++ FEBuilderGBA.Avalonia/ViewModels/SoundRoomViewerViewModel.cs
@@ -58,12 +58,12 @@ namespace FEBuilderGBA.Avalonia.ViewModels
             if (addr + dataSize > (uint)rom.Data.Length) return;
 
             CurrentAddr = addr;
-            SongId = rom.u16(addr + 0);
-            Raw4 = rom.u32(addr + 4);
-            Raw8 = rom.u32(addr + 8);
+            SongId = rom.u32(addr + 0);      // D0
+            Raw4 = rom.u32(addr + 4);        // D4
+            Raw8 = rom.u32(addr + 8);        // P8
             if (dataSize >= 16)
             {
-                TextId = rom.u16(addr + 12);
+                TextId = rom.u32(addr + 12); // D12
             }
             else
             {
